% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModels.R
\docType{package}
\name{hybridModels-package}
\alias{hybridModels-package}
\title{hybridModels: An R Package for the Stochastic Simulation of Disease Spreading in Dynamic Networks}
\description{
Simulates stochastic hybrid models for transmission of infectious diseases in dynamic networks. It is a metapopulation model in which each node in the network is a sub-population and disease spreads within nodes and among them, combining two approaches: stochastic simulation algorithm (\doi{10.1146/annurev.physchem.58.032806.104637}) and individual-based approach, respectively. Equations that models spread within nodes are customizable and there are two link types among nodes: migration and influence (commuting). More information in Fernando S. Marques, Jose H. H. Grisi-Filho, Marcos Amaku et al. (2020) \doi{10.18637/jss.v094.i06}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fernandosm/hybridModels}
  \item Report bugs at \url{https://github.com/fernandosm/hybridModels/issues}
}

}
\author{
\strong{Maintainer}: Fernando S. Marques \email{fernandosix@gmail.com}

Authors:
\itemize{
  \item Jose H. H. Grisi-Filho \email{zegrisi@gmail.com}
  \item Marcos Amaku \email{marcos.amaku@gmail.com}
}

}
\keyword{internal}
