% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModels.R
\docType{package}
\name{hybridModels}
\alias{hybridModels}
\title{hybridModels: an R package for stochastic simulation of disease spreading in dynamic networks.}
\description{
The hybridModels package provides functions to simulate stochastic models
in dynamics networks, using two processes in different scales:
1 Global scale to simulate the transmission from one node to another
2 Local scale to simulate the transmission inside the node
}
\section{Modeling transmission of diseases}{

'Framework to run n simulations in dynamic network and plot results'
}

