% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{networkSample}
\alias{networkSample}
\title{Daily record of animal's movement (from 2012 to 2013).}
\format{
A data frame with 78 rows and 4 variables:
\itemize{
  \item Day: The day when the movement occurs
  \item originID: The ID of the origin premises
  \item destinationID: The ID of the destination premises
  \item num.animals: The number of animals traded
}
}
\source{
ADAGRO
}
\usage{
networkSample
}
\description{
One dataset containing the number of animals that were moved from one node to 
another.
}
\keyword{datasets}
