% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{nodesCensus}
\alias{nodesCensus}
\title{Information about animal premises (from 2012 to 2013)}
\format{A data frame with 507 rows and 2 variables:
\describe{
  \item{nodes.ID}{The ID of the premises}
  \item{pop}{premises's population size}
}}
\source{
Hypothetical
}
\usage{
nodesCensus
}
\description{
A dataset containing animal premises' identification and census.
}
\keyword{datasets}

