% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{nodesInfo}
\alias{nodesInfo}
\title{Information about animal premises (from 2012 to 2013)}
\format{A data frame with 507 rows and 2 variables:
\describe{
  \item{nodes.ID}{The ID of the property}
  \item{pop}{Property's population size}
}}
\source{
Hypothetical
}
\usage{
nodesInfo
}
\description{
A dataset containing animal premises' identification and census.
}
\keyword{datasets}

