% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_model_class.R
\name{buildModelClass}
\alias{buildModelClass}
\title{It builds an object of a pre-specified class.}
\usage{
buildModelClass(x, var.names, init.cond, model.parms)
}
\arguments{
\item{x}{is an empty object of the will br built.}

\item{var.names}{a \code{\link{list}} with variable names of the network:
       the donor node, the reciever node, the time when each connection between
       donor to the reciever happened and the weight of these connection.
       The variables names must be "from", "to", "Time" and "arc", respectively.}

\item{init.cond}{a named \code{\link{vector}} with initial conditions.}

\item{model.parms}{a named \code{\link{vector}} with model's parameters.}
}
\value{
An object of the class requested.
}
\description{
\code{buildModelClass} is generic function that calls a method
              to create a object base on model's name.
}
\references{
.
}

