% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_model_class.R
\name{buildModelClass}
\alias{buildModelClass}
\title{It builds an object of a pre-specified class.}
\usage{
buildModelClass(x, var.names, init.cond, model.parms, prop.func = NULL,
  state.var = NULL, state.change.matrix = NULL)
}
\arguments{
\item{x}{is an empty object of a class requested.}

\item{var.names}{a \code{\link{list}} with variable names of the network:
       the donor node, the reciever node, the time when each connection between
       donor to the reciever happened and the weight of these connection.
       The variables names must be "from", "to", "Time" and "arc", respectively.}

\item{init.cond}{a named \code{\link{vector}} with initial conditions.}

\item{model.parms}{a named \code{\link{vector}} with model's parameters.}

\item{prop.func}{a character \code{\link{vector}} with propensity functions
       of a generic node. See references for more details}

\item{state.var}{a character \code{\link{vector}} with the state varialbes of
       the propensity functions.}

\item{state.change.matrix}{is a state-change \code{\link{matrix}}. See references
       for more details}
}
\value{
An object of the class requested.
}
\description{
\code{buildModelClass} is generic function that calls a method
              to create a object base on model's name.
}
\references{
.
}

