% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hybrid_Model.R
\name{hybridModel}
\alias{hybridModel}
\title{Hybrid model simulation.}
\usage{
hybridModel(network, var.names, nodesCensus, model.parms,
  model = "SI model without demographics", sim.number = 1,
  init.cond = init.cond, pop.correc = TRUE, num.cores = "max",
  ssa.method = list(method = character(), tau = integer()))
}
\arguments{
\item{network}{a \code{\link{data.frame}} with variables that describe
       the donor node, the reciever node, the time when each connection between
       donor to the reciever happened and the weight of these connection.}

\item{var.names}{a \code{\link{list}} with variable names of the network:
       the donor node, the reciever node, the time when each connection between
       donor to the reciever happened and the weight of these connection.
       The variables names must be "from", "to", "Time" and "arc", respectively.}

\item{nodesCensus}{a \code{\link{data.frame}} with the first column describing
       nodes' ID, the second column with the number of individuals and the third
       describing the day of the census (this last variable is not necessary
       for all models).}

\item{model.parms}{a named \code{\link{vector}} with model's parameters.}

\item{model}{a \code{\link{character}} describing model's name.}

\item{sim.number}{Number of repetitions.The default value is 1}

\item{init.cond}{a named \code{\link{vector}} with initial conditions.}

\item{pop.correc}{Whether \code{hybridModel} function tries to balance the number
       of individuals or not. The default value is TRUE.}

\item{num.cores}{number of  threads/cores that the simulation will use. the
       default value is num.cores = 'max', the algothim will use all
       threads/cores available.}

\item{ssa.method}{a \code{\link{list}} with ssa method and tau leap time. The
       default method is the direct method.}
}
\value{
Object containing a \code{\link{data.frame}} (results) with the number
        of individuals through time per node and per state.
}
\description{
\code{hybridModel} function runs hybrid models simulations.
}
\examples{
# Parameters and initial conditions.
data(networkSample, nodesCensus)
networkSample <- networkSample[which(networkSample$Dia < "2012-02-01"),]
var.names <- list(from = 'originID', to = 'destinyID', Time = 'Dia',
                  arc = 'num.animais')
model.parms <- c(Beta = 1e-4)
init.cond <- c(I100525 = 10, I1155 = 10, I100324 = 10)

# running simulations
sim.results <- hybridModel(network = networkSample, var.names,
                           nodesCensus = nodesCensus,
                           model.parms = model.parms,
                           model = 'SI model without demographics',
                           sim.number = 1,
                           init.cond = init.cond, num.cores = 1)

plot(sim.results)
}
\references{
.
}
\seealso{
\link{GillespieSSA}.
}

