% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridModels.R
\docType{package}
\name{hybridModels}
\alias{hybridModels}
\alias{hybridModels-package}
\title{hybridModels: A package for simulating hybrid models in dynamics networks.}
\description{
The hybridModels package provides functions to simulate stochastics models
in dynamics networks, using two processes in different scales:
1 Global scale to simulate the transmission from one node to another
2 Local sacale to simulate the transmission inside the node
}
\section{The models are}{

'SI model without demographics'
}

