% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{networkSample}
\alias{networkSample}
\title{Daily record of animal's movement (from 2012 to 2013)}
\format{A data frame with 31904 rows and 5 variables:
\describe{
  \item{Finalidade}{Type of destiny premises}
  \item{Dia}{The day when the movement occurs}
  \item{originID}{The ID of the origin premises}
  \item{destinyID}{The ID of the destiny premises}
  \item{num.animais}{The number of animals traded}
}}
\source{
ADAGRO
}
\usage{
networkSample
}
\description{
One dataset containing the number of animals that were moved from one node to
another.
}
\keyword{datasets}

