% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HM.R
\name{plot}
\alias{plot}
\alias{plot.HM}
\title{Plot for SI without demographics model}
\usage{
\method{plot}{HM}(x, sim = 1, plot.type = "subpop",
  facet.scales = "free_y", ...)
}
\arguments{
\item{x}{\code{HM} object}

\item{sim}{points to which simulation to plot.}

\item{plot.type}{plots the mean number of each state variable for the whole
population ('pop.mean'), or the subpopulations of a particular
simulation ('subpop'), or the mean of each subpopulation ('subpop.mean').}

\item{facet.scales}{should scales be fixed ("free_y", the default), free ("free"), or free
in one dimension ("free_x", "free_y"). See ggplot2 package for more
details.}

\item{...}{arguments to be passed to methods.}
}
\description{
\code{plot.HM} is a method to plot hybrid models from this
              package
}
\examples{
# Parameters and initial conditions.
data(networkSample)
networkSample <- networkSample[which(networkSample$Dia < "2012-06-30"),]

var.names <- list(from = 'originID', to = 'destinyID', Time = 'Dia',
                  arc = 'num.animais')
                  
prop.func <- c('beta * S * I / (S + I + R)', 'gamma * I')
state.var <- c('S', 'I', 'R')
state.change.matrix <- matrix(c(-1,  0, # S
                                 1, -1, # I
                                 0,  1), # R
                              nrow = 3, ncol = 2, byrow = TRUE)
                              
model.parms <- c(beta = 0.1, gamma = 0.01)

init.cond <- rep(200, length(unique(c(networkSample$originID,
                                      networkSample$destinyID))))
names(init.cond) <- paste('S', unique(c(networkSample$originID,
                                        networkSample$destinyID)), sep = '')
init.cond <- c(init.cond, c(I36811 = 50, I36812 = 50)) # adding infection
                  
# running simulations 
sim.results <- hybridModel(network = networkSample, var.names = var.names,
                           model.parms = model.parms, state.var = state.var,
                           prop.func = prop.func, init.cond = init.cond,
                           state.change.matrix = state.change.matrix,
                           sim.number = 2, num.cores = 'max')

# default plot layout (plot.types: 'pop.mean', 'subpop', or 'subpop.mean')
plot(sim.results, plot.type = 'subpop.mean')

# changing plot layout with ggplot2 (example)
library(ggplot2)
plot(sim.results, plot.type = 'subpop') + ggtitle('New Layout') + 
   theme(axis.title = element_text(size = 14, face = "italic"))
   
}

