% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_HM.R
\name{plot}
\alias{plot}
\alias{plot.HM}
\title{Plot for SI without demographics model}
\usage{
\method{plot}{HM}(x, sim = 1, ...)
}
\arguments{
\item{x}{\code{HM} object}

\item{sim}{which simulation to plot.}

\item{...}{arguments to be passed to methods}
}
\description{
\code{plot.HM} is a method to plot hybrid models from this
              package
}
\examples{
# Parameters and initial conditions.
data(networkSample, nodesCensus)
networkSample <- networkSample[which(networkSample$Dia < "2012-01-25"),]
var.names <- list(from = 'originID', to = 'destinyID', Time = 'Dia',
                  arc = 'num.animais')
model.parms <- c(Beta = 1e-4)
init.cond <- c(I100525 = 10, I1155 = 10, I100324 = 10)

# running simulations
sim.results <- hybridModel(network = networkSample, var.names,
                           nodesCensus = nodesCensus,
                           model.parms = model.parms,
                           model = 'SI model without demographics',
                           sim.number = 1, init.cond = init.cond,
                           num.cores = 1)

plot(sim.results)
}

