% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_siWoDemogr.R
\name{plot}
\alias{plot}
\alias{plot.siWoDemogr}
\title{Plot for SI without demographics model}
\usage{
\method{plot}{siWoDemogr}(x, sim = 1, ...)
}
\arguments{
\item{x}{\code{siWoDemogr} object}

\item{sim}{which simulation to plot.}

\item{...}{arguments to be passed to methods}
}
\description{
\code{plot.siWoDemogr} is a method to plot the SI without
              demographics model
}
\examples{
# Parameters and initial conditions.
data(networkSample, nodesInfo)
networkSample <- networkSample[which(networkSample$Dia < "2012-02-01"),]
var.names <- list(from = 'originID', to = 'destinyID', Time = 'Dia',
                  arc = 'num.animais')
model.parms <- c(Beta = 1e-4)
init.cond <- c(I100525 = 10, I1155 = 10, I100324 = 10)

# running simulations
sim.results <- hybridModel(network = networkSample, var.names,
                           nodes.info = nodesInfo,
                           model.parms = model.parms,
                           model = 'SI model without demographics',
                           sim.number = 2, init.cond = init.cond,
                           num.cores = 2)

plot(sim.results)
}

