% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate_HM.R
\name{simHM}
\alias{simHM}
\title{It runs the chosen hybrid model.}
\usage{
simHM(x, network, sim.number, num.cores = "max")
}
\arguments{
\item{x}{of a specific class of model.}

\item{network}{a \code{\link{data.frame}} with variables that describe
       the donor node, the reciever node, the time when each connection between
       donor to the reciever happened and the weight of these connection.}

\item{sim.number}{Number of repetitions.The default value is 1}

\item{num.cores}{number of  threads/cores that the simulation will use. the
       default value is num.cores = 'max', the algothim will use all
       threads/cores available.}
}
\value{
A \code{\link{data.frame}} with the number of individuals through
        time per node, per state and per simulation.
}
\description{
\code{simHM} is generic function that calls a method to run the
              simulation base on object's class
}
\references{
.
}
\seealso{
\link{GillespieSSA}.
}

